fix: only lock semaphore during active queries to prevent deadlocks
